Refactor packet manager

+++ src/ips_options/ips_icmp_id.cc
@@ -146,7 +146,7 @@ int IcmpIdOption::eval(Packet *p)
     PREPROC_PROFILE_START(icmpIdPerfStats);
 
     if( (p->icmph->type == ICMP_ECHO || p->icmph->type == ICMP_ECHOREPLY)
-        || (p->icmph->type == ICMP6_ECHO || p->icmph->type == ICMP6_REPLY)
+        || (p->icmph->type == (uint16_t)ICMP6_ECHO || p->icmph->type == (uint16_t)ICMP6_REPLY)
       )
     {
         /* test the rule ID value against the ICMP extension ID field */
